tests for signature and signature expiration
